support both global and local indexes
